requirements.txt
python-dotenv
pandas
google-generativeai
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2